Moves component renderers # Description - Moves component renderers to ui-components
