+++ src/ui-components/summary/summary.renderer.ts
@@ -1,7 +1,7 @@
 import * as d3 from 'd3';
-import * as expr from './expr';
-import { RenderInfo } from './models/render-info';
-import { Summary } from './ui-components/summary/summary.model';
+import * as expr from '../../expr';
+import { RenderInfo } from '../../models/render-info';
+import { Summary } from './summary.model';
 
 // TODO ???
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
